back-end/requirements.txt
Django==2.2.4
django-cors-headers==3.1.1
djangorestframework==3.10.2
psycopg2==2.9.3
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
